Rename AlchemySolidus::TouchAlchemyIngredients
If our namespace should be Alchemy::Solidus, then let's rename this
file, too.
